FIX: move use case models out of itwinai
