+++ use-cases/zebra2horse/cyclegan.py
