+++ use-cases/zebra2horse/pix2pix.py
